Migrate Java 9 test out of supersource
